refactor: 冗長なpackageを削除した
