--- lib/camera/camera.go
@@ -1,30 +1,0 @@
-package camera
-
-import (
-	"github.com/hajimehoshi/ebiten/v2"
-	gc "github.com/kijimaD/ruins/lib/components"
-	w "github.com/kijimaD/ruins/lib/world"
-	ecs "github.com/x-hgg-x/goecs/v2"
-)
-
-// SetTranslate はカメラを考慮した画像配置オプションをセットする
-// TODO: ズーム率を追加する
-func SetTranslate(world w.World, op *ebiten.DrawImageOptions) {
-	var camera *gc.Camera
-	var cPos *gc.Position
-	world.Manager.Join(
-		world.Components.Camera,
-		world.Components.Position,
-	).Visit(ecs.Visit(func(entity ecs.Entity) {
-		camera = world.Components.Camera.Get(entity).(*gc.Camera)
-		cPos = world.Components.Position.Get(entity).(*gc.Position)
-	}))
-
-	cx, cy := float64(world.Resources.ScreenDimensions.Width/2), float64(world.Resources.ScreenDimensions.Height/2)
-
-	// カメラ位置
-	op.GeoM.Translate(float64(-cPos.X), float64(-cPos.Y))
-	op.GeoM.Scale(camera.Scale, camera.Scale)
-	// 画面の中央
-	op.GeoM.Translate(float64(cx), float64(cy))
-}
