--- lib/camera/doc.go
@@ -1,2 +1,0 @@
-// Package camera はカメラ機能に関連するユーティリティを提供する。
-package camera
